refactor(broker): extract interface for command API service
